requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu130/
paddlepaddle-gpu==3.3.0
paddleocr[doc-parser]
https://paddle-whl.bj.bcebos.com/nightly/cu126/safetensors/safetensors-0.6.2.dev0-cp38-abi3-linux_x86_64.whl
fastapi
uvicorn
python-multipart
openai
python-dotenv